requirements.txt
python-telegram-bot
yfinance
apscheduler
flask